Moved Mainfeed.jsx back to original folder and created index.js in MainFeed folder. Will migrate once merged with team. Experimenting with Luxon for time remaining. Able to get time left in months, days, hours, minutes, and seconds.
