+++ client/src/components/MainFeed.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Feed from './MainFeed/index';
+// import axios from 'axios';
+
+const MainFeed = () => (
+  <div>
+    <Feed />
+  </div>
+);
+
+export default MainFeed;
