+++ client/testData/testData.js
@@ -6,13 +6,13 @@ export default {
       helper_id: null,
       address_id: 1,
       description: 'I need someone to come check on my dogs once a day for the next three days. They are very friendly. Two small poodles, hypoallergenic, about 20 pounds each. Just need somone to make sure their water bowls are filled. Thank you guys!',
-      car_required: false,
+      car_required: true,
       status: false,
       category: 'sitter',
       date: '2021-07-20',
       time: '12:21:13',
       date_requested: '2021-07-15',
-      duration: null,
+      duration: 'short: < 1 hour',
       user: {
         id: 1,
         firstname: 'Huge',
