+++ package.json
@@ -24,6 +24,7 @@
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
+    "luxon": "^2.0.1",
     "morgan": "^1.10.0",
     "parcel-bundler": "^1.12.5",
     "pg": "^8.6.0",
